Create `reverse_argparse` module (#1873)
Create a module that is able to generate the effective command line
invocation of a script, given the `ArgumentParser` that was used to
parse the command line options, and the `Namespace` of those parsed
options.  This makes it such that users can know exactly what was run,
including all default values, and any transformations that might've been
made to the arguments after parsing.

This commit only implements the functionality we care about at the
moment, which includes the following "actions":
* store
* store_true
* store_false
* append
* help
* version

The remaining actions are stubbed out but not yet implemented:
* store_const
* append_const
* count
* extend
* BooleanOptionalAction

Also how sub-parsers are handled remains to be determined in the future.