fix: Omit suppressed arguments
If an argument's help text has been suppressed, and the value of the
argument matches the default value, that indicated that a parser author
has hidden an argument from users, and the user hasn't modified it on
the command line.  To match the parser author's intent, such arguments
should be omitted from the effective command line invocation.