feat: Make private method public
Since classes/scripts using `reverse_argparse` may also need the ability
to quote a command line argument if there are spaces in it, transition
`quote_arg_if_necessary` from a private to a public method.  This will
avoid code duplication outside `reverse_argparse`, and ensure classes
and scripts that are using it are quoting arguments consistently.