test: Rework tests for calling program name
As part of the transition from GitLab CI/CD to GitHub Actions, the
program name used when running the tests has shifted from `__main__.py`
to `pytest`.  This commits adjusts the tests to ignore the program name,
and only pay attention to the arguments that come after it, such that
the tests will work in both contexts.