patch: Force `prefer_short` to be keyword-only
Turning on `flake8-boolean-trap` linting via Ruff resulted in the
following findings:

    reverse_argparse/reverse_argparse.py:242:31:
    FBT002 Boolean default positional argument in function definition
    reverse_argparse/reverse_argparse.py:242:31:
    FBT001 Boolean-typed positional argument in function definition
    Found 2 errors.

Switching `prefer_short` from a positional to a keyword-only argument
addresses the problem.

Note that this is technically a breaking change, but only for a
"private" method, not in the package's public API.  Therefore the change
is not registered as a breaking change via Conventional Commit syntax,
and no major version update will be created.  Instead, this commit will
force the creation of a patch release.  If users were relying on the
prior behavior of this internal method, they can simply switch to the
keyword syntax when calling it.