ci: Don't fix Ruff issues
Remove the `--fix` argument from the `ruff` check in the pre-commit
configuration.  We don't want automated commits pushed to PRs, because
we want to encourage contributors to clean up their branches before
merging.  Additionally, when pre-commit doesn't automatically fix files,
it shows the user what the issues are, thereby training them to avoid
them in the future.