docs: Fix highlighting in examples
I never noticed when we added the license and copyright information to
the top of all the source files that we didn't account for the
lines highlighted in the included examples.  This adjusts things such
that (1) we don't include all that header stuff when generating the
docs, and only include the lines of code that are really relevant to
share on the examples page, and (2) the appropriate lines are
highlighted.