ci: Subdivide release job
This is needed because the SLSA provenance reusable workflow cannot be
used as a step within a job, but must be used as a job on its own.  This
commit therefore subdivides the `release` job into a `release` job,
which runs `python-semantic-release` to create a new release, if
applicable, and then a `publish` job, to publish the release to PyPI and
GitHub Releases, if one was created.  We'll then be able to insert the
SLSA provenance job between the two.