fix: Rework release publishing
The PyPI upload failed with:

    Checking dist/reverse_argparse-2.0.8-py3-none-any.whl: PASSED
    Checking dist/multiple.intoto.jsonl: ERROR    InvalidDistribution:
    Unknown distribution format:
             'multiple.intoto.jsonl'

I suppose the provenance can't be included in the `dist` directory for
the PyPI upload, but it needs to be in there for python-semantic-release
publish to GitHub Releases action.  This commit reorders things, such
that we publish to GitHub first, and save PyPI for last, and between the
two we remove the provenance from the `dist` directory.  Hopefully this
works.